Restructurize tests and update cache busting
