+++ package.json
@@ -14,12 +14,13 @@
   "author": "Romain Dardour <romain@hull.io>",
   "license": "MIT",
   "scripts": {
-    "test": "npm run test:lint && npm run test:flow && npm run test:units",
+    "test": "npm run test:lint && npm run test:flow && npm run test:unit && npm run test:integration",
     "test:lint": "eslint src",
     "test:modules": "npm outdated --depth=0",
-    "test:units": "NODE_ENV=test mocha --require babel-register -R spec ./tests/*.js ./tests/**/*.js ./spec/*.js",
+    "test:unit": "NODE_ENV=test mocha --require babel-register -R spec ./test/unit/*.js ./test/unit/**/*.js",
+    "test:integration": "NODE_ENV=test mocha --require babel-register -R spec ./test/integration/*.js",
     "test:flow": "flow check",
-    "test:coverage": "nyc --babel-cache --hook-run-in-context false npm run test:units",
+    "test:coverage": "nyc --babel-cache --hook-run-in-context false npm run test",
     "update": "updtr",
     "clean": "rimraf lib",
     "build": "gulp build",
