+++ test/unit/support/hull-stub.js
